+++ frontend/src/index.ts
@@ -8,6 +8,7 @@ import "./shoelace";
 import { LocalePicker } from "./components/locale-picker";
 import { Alert } from "./components/alert";
 import { AccountSettings } from "./components/account-settings";
+import { InviteForm } from "./components/invite-form";
 import { SignUp } from "./pages/sign-up";
 import { Verify } from "./pages/verify";
 import { LogInPage } from "./pages/log-in";
@@ -516,4 +517,5 @@ customElements.define("btrix-archives", Archives);
 customElements.define("btrix-archive", Archive);
 customElements.define("btrix-archive-configs", ArchiveConfigsPage);
 customElements.define("btrix-account-settings", AccountSettings);
+customElements.define("btrix-invite-form", InviteForm);
 customElements.define("btrix-reset-password", ResetPassword);
