+++ frontend/src/pages/archive.ts
@@ -5,7 +5,7 @@ import type { AuthState, CurrentUser } from "../types/auth";
 import type { ArchiveData } from "../utils/archives";
 import LiteElement, { html } from "../utils/LiteElement";
 import { needLogin } from "../utils/auth";
-import { isOwner, AccessCode } from "../utils/archives";
+import { isOwner } from "../utils/archives";
 
 export type ArchiveTab = "settings" | "members";
 
@@ -31,12 +31,6 @@ export class Archive extends LiteElement {
 
   @state()
   archive?: ArchiveData;
-
-  @state()
-  isSubmitting: boolean = false;
-
-  @state()
-  serverError?: string;
 
   @state()
   successfullyInvitedEmail?: string;
@@ -87,7 +81,7 @@ export class Archive extends LiteElement {
           <sl-tab-panel
             name="settings"
             ?active=${this.archiveTab === "settings"}
-            >${this.renderSetting()}</sl-tab-panel
+            >${this.renderSettings()}</sl-tab-panel
           >
           <sl-tab-panel name="members" ?active=${this.archiveTab === "members"}>
             ${this.isAddingMember
@@ -99,7 +93,7 @@ export class Archive extends LiteElement {
     </article>`;
   }
 
-  private renderSetting() {
+  private renderSettings() {
     return html` TODO `;
   }
 
@@ -130,7 +124,7 @@ export class Archive extends LiteElement {
               <div class="border-b flex" role="row">
                 <div class="w-1/2 p-3" role="cell">(TODO) ${id}</div>
                 <div class="p-3" role="cell">
-                  ${isOwner(accessCode) ? msg("Owner") : msg("Member")}
+                  ${isOwner(accessCode) ? msg("Admin") : msg("Viewer")}
                 </div>
               </div>
             `
@@ -140,7 +134,7 @@ export class Archive extends LiteElement {
   }
 
   private renderAddMember() {
-    let successMessage, formError;
+    let successMessage;
 
     if (this.successfullyInvitedEmail) {
       successMessage = html`
@@ -150,14 +144,6 @@ export class Archive extends LiteElement {
               str`Successfully invited ${this.successfullyInvitedEmail}`
             )}</bt-alert
           >
-        </div>
-      `;
-    }
-
-    if (this.serverError) {
-      formError = html`
-        <div class="mb-5">
-          <bt-alert id="formError" type="danger">${this.serverError}</bt-alert>
         </div>
       `;
     }
@@ -176,52 +162,12 @@ export class Archive extends LiteElement {
 
       <div class="mt-3 border rounded-lg p-4 md:p-8 md:pt-6">
         <h2 class="text-lg font-medium mb-4">${msg("Add New Member")}</h2>
-
-        <sl-form
-          class="max-w-md"
-          @sl-submit=${this.onSubmitInvite}
-          aria-describedby="formError"
-        >
-          <div class="mb-5">
-            <sl-input
-              id="inviteEmail"
-              name="inviteEmail"
-              type="email"
-              label="${msg("Email")}"
-              placeholder="team-member@email.com"
-              required
-            >
-            </sl-input>
-          </div>
-          <div class="mb-5">
-            <sl-radio-group label="Select an option">
-              <sl-radio name="role" value=${AccessCode.owner}>
-                ${msg("Admin")}
-                <span class="text-gray-500">
-                  - ${msg("Can view, run, configure and manage crawls")}</span
-                >
-              </sl-radio>
-              <sl-radio name="role" value=${AccessCode.viewer} checked>
-                ${msg("Viewer")}
-                <span class="text-gray-500">
-                  - ${msg("Can only view crawls")}</span
-                >
-              </sl-radio>
-            </sl-radio-group>
-          </div>
-
-          ${formError}
-
-          <div>
-            <sl-button type="primary" submit>${msg("Invite")}</sl-button>
-            <sl-button
-              type="text"
-              href=${`/archives/${this.archiveId}/members`}
-              @click=${this.navLink}
-              >${msg("Cancel")}</sl-button
-            >
-          </div>
-        </sl-form>
+        <btrix-invite-form
+          @success=${this.onInviteSuccess}
+          @cancel=${() => this.navTo(`/archives/${this.archiveId}/members`)}
+          .authState=${this.authState}
+          .archiveId=${this.archiveId}
+        ></btrix-invite-form>
       </div>
     `;
   }
@@ -232,37 +178,8 @@ export class Archive extends LiteElement {
     return data;
   }
 
-  async onSubmitInvite(event: { detail: { formData: FormData } }) {
-    if (!this.authState) return;
-
-    this.isSubmitting = true;
-
-    const { formData } = event.detail;
-    const inviteEmail = formData.get("inviteEmail") as string;
-
-    try {
-      await this.apiFetch(
-        `/archives/${this.archiveId}/invite`,
-        this.authState,
-        {
-          method: "POST",
-          body: JSON.stringify({
-            email: inviteEmail,
-            role: Number(formData.get("role")),
-          }),
-        }
-      );
-
-      this.successfullyInvitedEmail = inviteEmail;
-    } catch (e: any) {
-      if (e?.isApiError) {
-        this.serverError = e?.message;
-      } else {
-        this.serverError = msg("Something unexpected went wrong");
-      }
-    }
-
-    this.isSubmitting = false;
+  onInviteSuccess(event: CustomEvent<{ inviteEmail: string }>) {
+    this.successfullyInvitedEmail = event.detail.inviteEmail;
   }
 
   updateUrl(event: CustomEvent<{ name: ArchiveTab }>) {
